Moved more generic functions to parent class Substrate
