+++ ContactMechanics/Systems.py
@@ -955,12 +955,13 @@ class NonSmoothContactSystem(SystemBase):
                 self.dual_hessian_product, x0=init_force, gtol=gtol,
                 maxiter=maxiter)
         elif solver == 'l-bfgs-b':
+            print("TEST")
             result = optim.minimize(
                 self.dual_objective(offset, gradient=True, logger=logger),
                 self.shape_minimisation_input(self.init_force),
                 method='L-BFGS-B', jac=True,
                 bounds=bnds,
-                options=dict(gtol=gtol, ftol=1e-20))
+                options=dict(gtol=gtol, ftol=1e-40))
 
         if result.success:
             # TODO: I think I need to call substrate.compute , so that substrate.energy is computed
